Only draw graphics when changes occur.
Added checks to only draw in the render window if
something in the graphics have changed. Otherwise,
it was always redrawing the same scene as fast as
it could (but below the maximum frame rate). This
change greatly improves responsiveness, and helps
save some CPU time because it is no longer constantly
drawing.